+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Review.cshtml.cs
@@ -124,12 +124,17 @@ namespace APIViewWeb.Pages.Assemblies
                 }
             }
 
-            if (ReviewContent.Directive == ReviewContentModelDirective.ErrorDueToInvalidAPIRevison)
+            if (ReviewContent.Directive == ReviewContentModelDirective.ErrorDueToInvalidAPIRevisonProceedWithPageLoad)
             {
                 NotificationMessage = ReviewContent.NotificationMessage;
             }
 
-            if (!ReviewContent.APIRevisionsGrouped.Any())
+            if (ReviewContent.Directive == ReviewContentModelDirective.ErrorDueToInvalidAPIRevisonRedirectToIndexPage)
+            {
+                return RedirectToPage("Index", new { notificationMessage = ReviewContent.NotificationMessage });
+            }
+
+            if (ReviewContent.APIRevisionsGrouped == null || !ReviewContent.APIRevisionsGrouped.Any())
             {
                 return RedirectToPage("LegacyReview", new { id = id });
             }