+++ tools/tsp-client/src/options.ts
@@ -97,7 +97,7 @@ export async function getOptions(): Promise<Options> {
     process.exit(1);
   }
 
-  let isUrl = false;
+  let isUrl = true;
   if (command === "init") {
     if (!values["tsp-config"]) {
       Logger.error("tspConfig is required");
@@ -105,7 +105,7 @@ export async function getOptions(): Promise<Options> {
       process.exit(1);
     }
     if (await doesFileExist(values["tsp-config"])) {
-      isUrl = true;
+      isUrl = false;
     }
     if (!isUrl) {
       if (!values.commit || !values.repo) {