+++ tools/pylint-extensions/azure-pylint-guidelines-checker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 0.3.1 (2023-1-16)
+
+- Docstring bug fix where paramtype was being considered for params
+
 ## 0.3.0 (2023-12-15)
 
 - Breaking changes involved in bump to pylint 3.0 support and bug fix to `incorrect-naming-convention` checker